requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
idna==3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
python-decouple==3.4
django-heroku==0.3.1
django==3.2
djangorestframework==3.12.4
pylint-django==2.4.2
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
whitenoise==5.2.0
psycopg2==2.9.1
dj-database-url==0.5.0